remove dead file, rename good test
